# cmd/docs/swagger.yaml
basePath: /
definitions:
  domain.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  response.AppError:
    properties:
      message:
        description: Message is the error message that may be displayed to end users
        type: string
      meta:
        description: Meta is the error detail detail data
    type: object
  response.AppSuccess:
    properties:
      data: {}
      meta: {}
    type: object
info:
  contact:
    email: zona.budi11@gmail.com
    name: API Support
  description: Only Test Purpose, If you Not Authorized, Please Login and input into
    security header with format Bearer token
  title: post App Majo Test APIs
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: get credential
      parameters:
      - description: Login Account
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppSuccess'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.AppError'
      summary: Login
      tags:
      - accounts
  /report-merchant:
    get:
      consumes:
      - application/json
      description: get list transaction merchant
      parameters:
      - description: page filter
        in: query
        minimum: 1
        name: page
        type: integer
      - description: limit data filter
        in: query
        minimum: 5
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppSuccess'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.AppError'
      security:
      - Bearer: []
      summary: List Transaction Merchant
      tags:
      - transactions
  /report-outlet:
    get:
      consumes:
      - application/json
      description: get list transaction Outlet
      parameters:
      - description: page filter
        in: query
        minimum: 1
        name: page
        type: integer
      - description: limit data filter
        in: query
        minimum: 5
        name: limit
        type: integer
      - description: Outlet ID
        in: path
        minimum: 1
        name: outlet_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AppSuccess'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.AppError'
      security:
      - Bearer: []
      summary: List Transaction Outlet
      tags:
      - transactions
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"